[TC-75]-refactor: movieName NOT_REQUIRED로 변경
