+++ src/main/java/com/server/booyoungee/domain/place/dto/response/PlaceSummaryResponse.java
@@ -1,19 +1,18 @@
-package com.server.booyoungee.domain.place.dto.response.movie;
+package com.server.booyoungee.domain.place.dto.response;
 
+import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.NOT_REQUIRED;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import java.util.List;
 
-import com.server.booyoungee.domain.like.domain.Like;
 import com.server.booyoungee.domain.place.domain.movie.MoviePlace;
-import com.server.booyoungee.domain.review.comment.domain.Comment;
 import com.server.booyoungee.domain.review.stars.domain.Stars;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 @Builder
-public record MoviePlaceSummaryResponse(
+public record PlaceSummaryResponse(
 	@Schema(description = "장소 ID", example = "1", requiredMode = REQUIRED)
 	Long id,
 
@@ -32,12 +31,12 @@ public record MoviePlaceSummaryResponse(
 	@Schema(description = "리뷰 수", example = "11", requiredMode = REQUIRED)
 	int reviews,
 
-	@Schema(description = "영화 제목", example = "극비 수사", requiredMode = REQUIRED)
+	@Schema(description = "영화 제목", example = "극비 수사", requiredMode = NOT_REQUIRED)
 	String movieName
 ) {
 
-	public static MoviePlaceSummaryResponse of (MoviePlace moviePlace, List<Stars> stars, int likes, int reviews) {
-		return MoviePlaceSummaryResponse.builder()
+	public static PlaceSummaryResponse of (MoviePlace moviePlace, List<Stars> stars, int likes, int reviews) {
+		return PlaceSummaryResponse.builder()
 			.id(moviePlace.getId())
 			.name(moviePlace.getName())
 			.basicAddress(moviePlace.getBasicAddress())
